refactor(mqtt): move `authenticate` function to `mqtt` module
